Revert "fix IO"
This reverts commit f9be2a9adc22551035536fe337c71b0b2fe8108a.

+++ code/fnodes/math/FNodeIndexInfo.cs
@@ -13,7 +13,7 @@ public class FNodeIndexInfo : FNode
             {
             "Index", new FOutputInt(this, delegate() {
                 inputs["File"].Get();
-                return (int)Project.idxEval;
+                return Project.idxEval;
             })},
             {
             "Loop Number", new FOutputInt(this, delegate() {
